// packages/gameSchema/src/types.ts
export type IntegerOrInput =
  | number
  | { readonly $input: string }
  | { readonly $input: string; readonly ifTrue: number; readonly ifFalse: number }

export type InputValue = number | string | boolean
// ...
export interface MarkDiceOperation {
  readonly operator: '>' | '>=' | '<' | '<=' | '='
  readonly value: IntegerOrInput
  readonly flag: string
}

export interface CapOperation {
  readonly min?: IntegerOrInput
  readonly max?: IntegerOrInput
}

export interface ModifyOperation {
  readonly keepHighest?: IntegerOrInput
  readonly keepLowest?: IntegerOrInput
  readonly add?: IntegerOrInput
  readonly cap?: CapOperation
  readonly markDice?: MarkDiceOperation
  readonly keepMarked?: string
}
// ...
export type RollInput = Readonly<Record<string, InputValue>>